refactor: restructure CLI commands into separate packages
Split the monolithic commands.go into individual command packages under cmd/dodot/commands/. Each command now has its own directory containing:
- Command definition (e.g., adopt.go) that returns a cobra.Command
- Message constants (msgs.go) with command-specific messages
- Embedded text files for long descriptions and examples

Key changes:
- commands.go renamed to root.go and refactored to import commands
- msgs.go updated to re-export messages from command packages
- Each command package is self-contained and follows Go conventions
- Maintains backward compatibility through message re-exports
- All tests continue to pass (856 tests successful)
