+++ cmd/dodot/commands/addignore/addignore-example.txt
