+++ cmd/dodot/commands/addignore/addignore-long.txt
