+++ cmd/dodot/commands/init/init-example.txt
