+++ cmd/dodot/commands/init/init-long.txt
