+++ cmd/dodot/commands/snippet/snippet-example.txt
