+++ cmd/dodot/commands/status/status-example.txt
