+++ cmd/dodot/commands/topics/msgs.go
@@ -1,0 +1,7 @@
+package topics
+
+// Message constants
+const (
+	MsgShort = "List all topics or show help for a topic"
+	MsgLong  = "Display a list of all available help topics that provide additional documentation beyond command help."
+)
