--- cmd/dodot/msgs/adopt-example.txt
@@ -1,11 +1,0 @@
-# Adopt a single file into a pack
-dodot adopt git ~/.gitconfig
-
-# Adopt multiple files at once
-dodot adopt shell ~/.bashrc ~/.bash_profile ~/.bash_aliases
-
-# Adopt a configuration directory
-dodot adopt vim ~/.config/nvim
-
-# Force adoption even if destination exists
-dodot adopt --force tmux ~/.tmux.conf
