--- cmd/dodot/msgs/adopt-long.txt
@@ -1,18 +1,0 @@
-Adopt existing configuration files into a pack by moving them to your dotfiles
-repository and immediately creating symlinks back to their original locations.
-
-This command streamlines the process of bringing files under dodot's management:
-1. Moves the source file/directory to the specified pack
-2. Creates a symlink from the original location to the new location
-
-The adoption is atomic - the file is moved and symlinked in one operation,
-ensuring your configuration continues to work without interruption.
-
-Smart path handling:
-- Files from $HOME go to the pack root (e.g., ~/.gitconfig -> pack/gitconfig)
-- Files from XDG paths preserve their structure (e.g., ~/.config/app/config.toml -> pack/app/config.toml)
-
-Safety features:
-- Won't overwrite existing files in the pack (use --force to override)
-- Detects if a file is already managed by dodot and exits gracefully
-- Shows clear output for each operation performed
