--- cmd/dodot/msgs/completion-long.txt
@@ -1,27 +1,0 @@
-Generate a shell completion script for dodot.
-
-Shell completions allow you to use <TAB> to complete commands, flags, and pack names.
-
-To enable completions, you need to:
-1. Generate the completion script for your shell
-2. Source it in your shell configuration
-
-Bash:
-  $ dodot completion bash > ~/.local/share/bash-completion/completions/dodot
-  $ source ~/.bashrc
-
-Zsh:
-  $ dodot completion zsh > ~/.zfunc/_dodot
-  $ echo "fpath=(~/.zfunc $fpath)" >> ~/.zshrc
-  $ echo "autoload -Uz compinit && compinit" >> ~/.zshrc
-  $ source ~/.zshrc
-
-Fish:
-  $ dodot completion fish > ~/.config/fish/completions/dodot.fish
-  $ source ~/.config/fish/config.fish
-
-PowerShell:
-  $ dodot completion powershell | Out-String | Invoke-Expression
-
-Note: After installing completions, you may need to restart your shell
-or source your shell configuration file.
