--- cmd/dodot/msgs/deprovision-example.txt
@@ -1,14 +1,0 @@
-# Deprovision a specific pack
-dodot deprovision vim
-
-# Deprovision multiple packs
-dodot deprovision vim zsh homebrew
-
-# Deprovision all packs
-dodot deprovision
-
-# Preview what would be removed (dry run)
-dodot deprovision --dry-run
-
-# Deprovision specific packs with dry run
-dodot deprovision vim zsh --dry-run
