--- cmd/dodot/msgs/deprovision-long.txt
@@ -1,14 +1,0 @@
-The deprovision command removes provisioning state for specified packs,
-effectively uninstalling or clearing resources that were deployed by
-provisioning handlers.
-
-This command only affects provisioning handlers (homebrew, provision)
-and preserves all linking handler state (symlinks, path, shell).
-
-When run without any pack names, it will deprovision all packs.
-Use the --dry-run flag to preview what would be removed without
-making any changes.
-
-Note: This command removes dodot's state tracking for provisioning
-operations. The actual uninstallation of packages or resources depends
-on the specific handler implementation.
