--- cmd/dodot/msgs/provision-example.txt
@@ -1,11 +1,0 @@
-  # Provision all packs
-  dodot provision
-  
-  # Provision specific packs
-  dodot provision vim zsh
-  
-  # Force re-run of provisioning handlers
-  dodot provision --force
-  
-  # Dry run to preview changes
-  dodot provision --dry-run
