--- cmd/dodot/msgs/unlink-example.txt
@@ -1,11 +1,0 @@
-# Unlink the vim pack
-dodot unlink vim
-
-# Unlink multiple packs
-dodot unlink vim git zsh
-
-# Preview what would be removed (dry run)
-dodot unlink vim --dry-run
-
-# Unlink all packs
-dodot unlink
